gateway/requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.30.0
grpcio==1.64.1
grpcio-tools==1.64.1
prometheus-client==0.20.0
pyyaml==6.0.2
protobuf==5.26.1
